pyproject.toml
[tool.poetry]
name = "placeholder"
version = "0.1.0"
description = ""
authors = ["bekzod-fayzikuloff <fayzikuloff.b@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
django-jazzmin = "^2.5.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
pep8 = "^1.7.1"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
pytest-xdist = "^2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
exclude = [
    '^.py$',
]
